% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edgemat.R
\name{edgemat}
\alias{edgemat}
\title{Generate the edge matrix using R}
\usage{
edgemat(p, pr)
}
\arguments{
\item{p}{Variable dimension}

\item{pr}{Probability matrix}
}
\value{
a matrix consisting of edge information
}
\description{
Generate relationships between variables with a set probability
}
